feat: adding Comment model.
The Comment model uses ContentType so we can allow a model to be related to any number of models by just adding three attributes to a Model:
A ForeignKey field that points to a ContentType. Normally this is called content_type
A PositiveIntegerField that stores the primary key of the related object. Normally this is called object_id
A GenericForeignKey field, a special type of field that will look up the object from the other two new fields.